@@ -1,31 +1,33 @@
 module github.com/aspect-build/gazelle_kt
 
-go 1.23.2
+go 1.24.1
 
 require (
-	github.com/aspect-build/aspect-cli v1.202449.18
+	github.com/aspect-build/aspect-cli v1.202449.19-0.20250711215336-dd3563abbc5b
 	github.com/bazel-contrib/rules_jvm v0.27.0
-	github.com/bazelbuild/bazel-gazelle v0.44.0
+	github.com/bazelbuild/bazel-gazelle v0.44.1-0.20250707152251-85aa3a252a58
 	github.com/emirpasic/gods v1.18.1
 	github.com/rs/zerolog v1.34.0
-	github.com/smacker/go-tree-sitter v0.0.0-20240827094217-dd81d9e9be82
-	go.starlark.net v0.0.0-20240725214946-42030a7cedce
+	go.starlark.net v0.0.0-20250318223901-d9371fef63fe
 )
 
 require (
-	github.com/bazelbuild/buildtools v0.0.0-20250204160707-ad48c76ab9b5 // indirect
+	github.com/bazelbuild/buildtools v0.0.0-20250326091033-f79c8eafbddd // indirect
+	github.com/bmatcuk/doublestar/v4 v4.8.1 // indirect
 	github.com/go-git/gcfg v1.5.1-0.20230307220236-3a3c6141e376 // indirect
-	github.com/go-git/go-billy/v5 v5.5.0 // indirect
-	github.com/go-git/go-git/v5 v5.12.0 // indirect
+	github.com/go-git/go-billy/v5 v5.6.2 // indirect
+	github.com/go-git/go-git/v5 v5.14.0 // indirect
 	github.com/google/btree v1.1.3 // indirect
 	github.com/itchyny/gojq v0.12.16 // indirect
 	github.com/itchyny/timefmt-go v0.1.6 // indirect
 	github.com/jbenet/go-context v0.0.0-20150711004518-d14ea06fba99 // indirect
 	github.com/mattn/go-colorable v0.1.14 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
-	golang.org/x/mod v0.23.0 // indirect
-	golang.org/x/net v0.34.0 // indirect
-	golang.org/x/sys v0.30.0 // indirect
+	github.com/smacker/go-tree-sitter v0.0.0-20240827094217-dd81d9e9be82 // indirect
+	golang.org/x/mod v0.24.0 // indirect
+	golang.org/x/net v0.40.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
 	golang.org/x/tools/go/vcs v0.1.0-deprecated // indirect
+	gopkg.in/op/go-logging.v1 v1.0.0-20160211212156-b2cb9fa56473 // indirect
 	gopkg.in/warnings.v0 v0.1.2 // indirect
 )
